requirements.txt
# -i https://pypi.org/simple
aio-pika==9.5.4; python_version >= '3.9' and python_version < '4.0'
aiormq==6.8.1; python_version >= '3.8' and python_version < '4.0'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.7.0; python_version >= '3.9'
certifi==2024.12.14; python_version >= '3.6'
click==8.1.7; python_version >= '3.7'
dnspython==2.7.0; python_version >= '3.9'
email-validator==2.2.0
exceptiongroup==1.2.2; python_version >= '3.7'
fastapi[standard]==0.115.6; python_version >= '3.8'
fastapi-cli[standard]==0.0.7
greenlet==3.1.1; python_version < '3.13' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httptools==0.6.4
httpx==0.28.1
idna==3.10; python_version >= '3.6'
jinja2==3.1.4
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
mysqlclient==2.2.6; python_version >= '3.8'
pamqp==3.3.0; python_version >= '3.7'
passlib==1.7.4
propcache==0.2.1; python_version >= '3.9'
pydantic==2.10.4; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
rich==13.9.4; python_full_version >= '3.8.0'
rich-toolkit==0.12.0; python_version >= '3.8'
shellingham==1.5.4; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.36; python_version >= '3.7'
starlette==0.41.3; python_version >= '3.8'
typer==0.15.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn[standard]==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
yarl==1.18.3; python_version >= '3.9'
